# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rasheed - Finance Savings Agent",
...
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/transactions/quick-log": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "تسجيل سريع لمعاملة",
        "description": "المسار الأهم بالتطبيق، ويدعم مسارين: (1) الأيقونات: إرسال `category_id` يحفظ فورًا بدون أي معالجة ذكية. (2) النص الذكي: إرسال `note` بدون `category_id` يجعل محرك التصنيف يخمّن التصنيف تلقائيًا، ويرجع `ai_suggested=true` مع `suggestion_confidence`. يجب إرسال `category_id` أو `note` على الأقل.",
        "operationId": "quick_log_transaction_api_v1_transactions_quick_log_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionQuickLogCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionOut"
                }
              }
            }
          },
          "401": {
            "description": "التوكن ناقص أو غير صالح أو منتهي الصلاحية، أو المستخدم غير موجود/غير مفعّل.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/transactions": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "قائمة المعاملات",
        "description": "يرجّع آخر معاملات المستخدم (الأحدث أولًا). لمعاملات الدخل، `unallocated_amount` هو المبلغ الذي لم يُوزَّع بعد على أهداف.",
        "operationId": "list_transactions_api_v1_transactions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "عدد المعاملات المطلوب إرجاعها (الافتراضي 50)",
              "default": 50,
              "title": "Limit"
            },
            "description": "عدد المعاملات المطلوب إرجاعها (الافتراضي 50)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionOut"
                  },
                  "title": "Response List Transactions Api V1 Transactions Get"
                }
              }
            }
          },
          "401": {
            "description": "التوكن ناقص أو غير صالح أو منتهي الصلاحية، أو المستخدم غير موجود/غير مفعّل.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/scan-receipt": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "قراءة فاتورة من صورة (OCR)",
        "description": "يستقبل صورة فاتورة (multipart، الحقل `file`، بصيغة JPEG أو PNG أو WEBP وبحد أقصى 10 ميغابايت) فيقرأها Gemini Vision ويرجّع **مسودة** فقط: مبلغ وتصنيف مقترح وملاحظة. لا يحفظ أي شيء؛ يراجع المستخدم المسودة ثم يحفظها عبر `POST /transactions/quick-log`. القيمة `readable=false` تعني أن الصورة غير واضحة أو ليست فاتورة.",
        "operationId": "scan_receipt_api_v1_transactions_scan_receipt_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_scan_receipt_api_v1_transactions_scan_receipt_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptScanResult"
                }
              }
            }
          },
          "401": {
            "description": "التوكن ناقص أو غير صالح أو منتهي الصلاحية، أو المستخدم غير موجود/غير مفعّل.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "صيغة الصورة غير مدعومة، أو حجمها أكبر من 10 ميغابايت، أو الصورة فارغة.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/transactions/parse-sms": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "تحليل رسالة بنكية ملصوقة",
        "description": "يحلّل نص رسالة بنكية (نسخ ولصق يدوي، بدون أي صلاحية قراءة رسائل) ويرجّع **مسودة** (مبلغ، نوع، ملاحظة). لا يحفظ شيئًا. القيمة `parsed=false` تعني أن النص لا يطابق أي نمط مدعوم.",
        "operationId": "parse_sms_api_v1_transactions_parse_sms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmsParseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmsParseResult"
                }
              }
            }
          },
          "401": {
            "description": "التوكن ناقص أو غير صالح أو منتهي الصلاحية، أو المستخدم غير موجود/غير مفعّل.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/transactions/sms-import/preview": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "معاينة استيراد رسائل بنكية",
        "description": "يحلّل رسائل قُرئت من صندوق الوارد على الجهاز (بعد فلترة محلية، حتى 300 رسالة) ويرجّع المعاملات المكتشفة مع علامة `already_imported` للرسائل المستوردة سابقًا. لا يحفظ شيئًا.",
        "operationId": "preview_sms_import_api_v1_transactions_sms_import_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmsImportPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmsImportPreviewResponse"
                }
              }
            }
          },
          "401": {
            "description": "التوكن ناقص أو غير صالح أو منتهي الصلاحية، أو المستخدم غير موجود/غير مفعّل.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/transactions/sms-import/confirm": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "تأكيد استيراد الرسائل",
        "description": "ينشئ معاملات فعلية من الرسائل التي اختارها المستخدم. السيرفر يعيد تحليل نص كل رسالة بنفسه ولا يثق بأي مبلغ قادم من الموبايل، ويتجاهل الرسائل المستوردة مسبقًا.",
        "operationId": "confirm_sms_import_api_v1_transactions_sms_import_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmsImportConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TransactionOut"
                  },
                  "type": "array",
                  "title": "Response Confirm Sms Import Api V1 Transactions Sms Import Confirm Post"
                }
              }
            }
          },
          "401": {
            "description": "التوكن ناقص أو غير صالح أو منتهي الصلاحية، أو المستخدم غير موجود/غير مفعّل.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "بعض الرسائل استُوردت للتو بطلب آخر؛ حدّث القائمة وأعد المحاولة.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/transactions/{transaction_id}/allocate": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "توزيع دخل على الأهداف",
        "description": "يوزّع معاملة دخل واحدة على هدف أو أكثر بقرار صريح من المستخدم (لا يوجد توجيه تلقائي). يقبل توزيعًا جزئيًا؛ يبقى الباقي في `unallocated_amount` حتى يوزّعه المستخدم لاحقًا. يرجّع المعاملة بعد التحديث والأهداف التي تغيّرت.",
        "operationId": "allocate_income_api_v1_transactions__transaction_id__allocate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Transaction Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncomeAllocationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncomeAllocationResult"
                }
              }
            }
          },
          "401": {
            "description": "التوكن ناقص أو غير صالح أو منتهي الصلاحية، أو المستخدم غير موجود/غير مفعّل.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "المعاملة ليست دخلًا، أو هدف مكرر/غير موجود/غير نشط، أو المبلغ أكبر من المتاح.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "المعاملة غير موجودة.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agent/actions": {
      "get": {
        "tags": [
          "Agent"
        ],
        "summary": "الاقتراحات المعلّقة",
        "description": "يرجّع اقتراحات رشيد التي لم يرد عليها المستخدم بعد (`pending`)، ويعرضها الموبايل كبطاقات موافقة/رفض. محتوى `payload` يختلف حسب `action_type`.",
        "operationId": "list_pending_actions_api_v1_agent_actions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentActionOut"
                  },
                  "type": "array",
                  "title": "Response List Pending Actions Api V1 Agent Actions Get"
                }
              }
            }
          },
          "401": {
            "description": "التوكن ناقص أو غير صالح أو منتهي الصلاحية، أو المستخدم غير موجود/غير مفعّل.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_scan_receipt_api_v1_transactions_scan_receipt_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "صورة الفاتورة (JPEG أو PNG أو WEBP، حتى 10 ميغابايت)"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_scan_receipt_api_v1_transactions_scan_receipt_post"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "رسالة الخطأ (بالعربي)"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "شكل الخطأ القياسي الراجع من كل الـ Endpoints.",
        "examples": [
          {
            "detail": "الهدف غير موجود"
          }
        ]
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReceiptScanResult": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "category_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Name"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "readable": {
            "type": "boolean",
            "title": "Readable",
            "description": "false إذا الصورة غير واضحة أو ليست فاتورة"
          }
        },
        "type": "object",
        "required": [
          "amount",
          "category_id",
          "category_name",
          "note",
          "readable"
        ],
        "title": "ReceiptScanResult",
        "description": "نتيجة قراءة فاتورة — مسودة بس، ما بتنشئ Transaction. الموبايل يعبّي فيها\nشاشة التسجيل السريع مسبقًا، والمستخدم يراجعها/يعدّلها قبل ما يحفظ فعليًا.",
        "examples": [
          {
            "amount": 12.5,
            "category_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
            "category_name": "بقالة وسوبرماركت",
            "note": "سوبرماركت",
            "readable": true
          }
        ]
      },
...
      "SmsImportPreviewRequest": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/SmsMessageIn"
            },
            "type": "array",
            "maxItems": 300,
            "minItems": 1,
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "messages"
        ],
        "title": "SmsImportPreviewRequest",
        "examples": [
          {
            "messages": [
              {
                "body": "Successfully received 4.500 JOD from 0791234567 Current balance JOD5,000.000 JOD.",
                "received_at": "2026-10-06T14:41:00Z"
              }
            ]
          }
        ]
      },
...
      "SmsParseRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 3,
            "title": "Text",
            "description": "نص الرسالة البنكية كما وصلت (CliQ)"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "SmsParseRequest",
        "examples": [
          {
            "text": "Successfully received 4.500 JOD from 0791234567 Current balance JOD5,000.000 JOD."
          }
        ]
      },
      "SmsParseResult": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TransactionType"
              },
              {
                "type": "null"
              }
            ]
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "parsed": {
            "type": "boolean",
            "title": "Parsed",
            "description": "false إذا النص لا يطابق أي نمط رسالة بنكية مدعوم"
          }
        },
        "type": "object",
        "required": [
          "amount",
          "type",
          "note",
          "parsed"
        ],
        "title": "SmsParseResult",
        "description": "نتيجة تحليل رسالة بنكية — مسودة بس، نفس فلسفة ReceiptScanResult تمامًا:\nما بتنشئ أي Transaction، الموبايل يعبّي فيها شاشة التسجيل السريع\nوالمستخدم يراجعها ويحفظها بنفسه.",
        "examples": [
          {
            "amount": 4.5,
            "note": "تحويل CliQ من 0791234567",
            "parsed": true,
            "type": "income"
          }
        ]
      },
...
      "TransactionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "type": {
            "$ref": "#/components/schemas/TransactionType"
          },
          "source": {
            "$ref": "#/components/schemas/TransactionSource",
            "description": "مصدر المعاملة: `manual` أو `ocr` أو `sms` (و`open_banking` محجوز للمستقبل)"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CategoryOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "التصنيف، أو null إذا المعاملة غير مصنّفة"
          },
          "ai_suggested": {
            "type": "boolean",
            "title": "Ai Suggested",
            "default": false
          },
          "suggestion_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggestion Confidence"
          },
          "unallocated_amount": {
            "type": "number",
            "title": "Unallocated Amount",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "amount",
          "type",
          "source",
          "note",
          "occurred_at",
          "category"
        ],
        "title": "TransactionOut",
        "examples": [
          {
            "ai_suggested": false,
            "amount": 45,
            "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
            "note": "أجرة يوم",
            "occurred_at": "2026-10-06T14:41:00Z",
            "source": "manual",
            "type": "income",
            "unallocated_amount": 45
          }
        ]
      },
      "TransactionQuickLogCreate": {
        "properties": {
          "amount": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Amount",
            "description": "المبلغ بالدينار، لازم يكون أكبر من صفر"
          },
          "type": {
            "$ref": "#/components/schemas/TransactionType",
            "description": "`income` دخل أو `expense` مصروف"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id",
            "description": "معرّف تصنيف من `GET /categories` (مسار الأيقونات)"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Note",
            "description": "ملاحظة نصية؛ بدون `category_id` تُستخدم لتخمين التصنيف تلقائيًا"
          },
          "occurred_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At",
            "description": "وقت المعاملة؛ لو ما تحدد، بنعتبره الآن"
          }
        },
        "type": "object",
        "required": [
          "amount",
          "type"
        ],
        "title": "TransactionQuickLogCreate",
        "description": "يدعم المسارين سوا:\n- مسار الأيقونات: يوصل category_id مباشرة → حفظ فوري بدون أي معالجة ذكاء اصطناعي\n- مسار النص الذكي: يوصل note بدون category_id → محرك التصنيف (Rule-based اليوم، Gemini لاحقًا)\n  بيحاول يخمّن التصنيف تلقائيًا",
        "examples": [
          {
            "amount": 4.5,
            "category_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
            "type": "expense"
          },
          {
            "amount": 12,
            "note": "غدا مع الشباب",
            "type": "expense"
          }
        ]
      },
...
      "TransactionType": {
        "type": "string",
        "enum": [
          "income",
          "expense"
        ],
        "title": "TransactionType"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}